Introduce minimal modeling
This first introduces the concept of a world with foxes and rabbits. It
also provides a minimal visual representation as a string of emoji.

Note that elm-format wants to convert some emoji to UTF-8 code points
while leaving the literal emoji in for others. I'm not sure why this is
the case but it won't allow me to consistently enter either literal
emojis or code points, instead forcing me into this half-and-half
situation.