Configure build to work on non-root URLs
All the assets were required with absolute paths. This breaks when
trying to run the game when not at the root of a domain, such as happens
when uploading to itch.io.

I was seeing seeing garbled Cyrillic characters instead of
emoji when building locally and opening the file in a browser. This is a
bit weird because it works fine when building with the same command on
Netlify. Chrome must detect the encoding differently when reading a
local file. I was able to fix the issue by setting an explicit encoding
in the HTML file.