Foxes eat neighboring rabbit
If there is more than one rabbit nearby, eat the first one.

I expected to be able to use `CellGrid.transform` but that function
only allows you to modify the current cell and not the neighbors.
Instead I built a custom function on top of `CellGrid.foldl` using the
grid itself as the accumulator.

This gives me the ability to modify the grid as a whole at each step. It
also gives me access to the modified grid at each step rather than the
initial grid like `CellGrid.transform` would have done. This is useful
because, unlike Conway's game of life, I want the changes at each cell
to impact the next cells as I'm traversing rather than calculating the
entire generation at once.

It's important to make sure to update the accumulated grid at each step,
it's not good enough to return a modified fox or rabbit. I need to
modify their cell in the grid, otherwise the changes wont propagate. Am
I re-creating mutable memory???