Rabbits attempt to flee foxes
If a rabbit is next to a fox it will try to move to a non fox-adjacent
cell. If there are multiple, it will go to the first possible option.

Because foxes can't move (yet), the game quickly goes static over
the first few turns as foxes eat adjacent rabbits and rabbits move away,
leading to separate islands of foxes and rabbits.

On the implementation side of things, I was using `CellGrid.Position` a
lot and it was annoyingly verbose so I exposed it to be able to use it
without the namespace prefix. I don't have any other kinds of positions
(for now?) so it wasn't worth having the extra specificity.