Fix bugs in jxxcarlson/elm-cell-grid
The cell-grid library I was using had bugs in some underlying functions
causing the `neighbors` function I was relying on to break. I opened an
[issue on GitHub](https://github.com/jxxcarlson/elm-cell-grid/issues/8)
describing the problem.

To enable me to keep moving forward, this commit copies the
`CellGrid.elm` file from the package into this project and modifies it
to fix the issues.

The main fixes are making sure x/y and columns/rows are passed as
arguments in a consistent order, and adding a check for out-of-bounds
coordinates when attempting to `get` a value.