Introduce rabbit hunger
This doesn't impact the gameplay much since rabbits can eat anywhere as
long as they don't use their turn to move. In rare circumstances a
rabbit can be chased until it starves to death.

The real value of having rabbits also include the hunger mechanic is
that it will regulate birth rates once we introduce those.