Introduce rabbit reproduction
Rabbits will reproduce if it's safe and they have enough food. This
leads to a rabbit boom when the fox population declines.

Ideally, this would mean alternating fox and rabbit booms but foxes tend
to starve to death even when there are plenty of rabbits because they
all move over to the top-right corner. Once we have random movement we
should get more realistic behavior.