Extract `Cell` and `Grid` modules
By extracting `Grid`, I can more easily write other modules that operate
over the grid type. This sets up some of the graphing work I have coming
up next.

The `Grid` module contains functions that are specific to my game grid
`CellGrid Cell` but tries to avoid having game rules in it. Ideally this
module has functions for all the queries and operations I might want to
do on a grid without implementing the actual game rules which should be
a higher-level on top of the grid.

I've left some functions that are game-rule-y in the `Main` module for
now since they don't block the graphing work I want to do. These
functions will probably need to be untangled and extracted at some
point. They do a mix of `Grid` and `CellGrid` operations and ideally the
game-rule functions never deal with `CellGrid` directly.